+++ hardhat-config/external.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  external: {
+    contracts: [
+      {
+        artifacts: 'node_modules/hardhat-deploy/extendedArtifacts',
+      },
+    ],
+  },
+};